Added tooltip ability for structure properties Added float to packets Changed forest rate for different forest levels Changed lumbermill abilities Fixed troop crashing bug when stub was null and a get troop info came in Hopefully fixed scheduler bug Relaxed attack points. Will gain 1 attack point per unit killed. Changed battle turn to 15 seconds per hit Forest sends system message to player when it's depleted completely

+++ Game/Database/DbLoader.cs
@@ -193,7 +193,7 @@ namespace Game.Database {
             Global.Logger.Info("Loading forests...");
             using (var reader = dbManager.Select(Forest.DB_TABLE)) {
                 while (reader.Read()) {
-                    Forest forest = new Forest((byte)reader["level"], (int)reader["capacity"], (int)reader["rate"]) {
+                    Forest forest = new Forest((byte)reader["level"], (int)reader["capacity"], (float)reader["rate"]) {
                         DbPersisted = true,
                         X = (uint)reader["x"],
                         Y = (uint)reader["y"],
